Move `MonadLifxIO` to its own module
Allows us to expose internals for advanced use cases while keeping main module clean.
